pyproject.toml
[project]
name = "zundamon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pykakasi>=2.3.0",
    "requests>=2.33.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]
